refactor: split tdlib-client into modules per TS conventions
src/
  types.ts       — shared interfaces (FormattedMessage, RawTdMessage, etc.)
  tdlib.ts       — detectTdlibPath, TDLib init
  auth.ts        — AuthManager (auth state machine)
  format.ts      — extractText, MessageFormatter
  resolve.ts     — Resolver (user/sender resolution)
  updates.ts     — UpdateDispatcher (TDLib update handlers)
  client.ts      — TelegramLSPClient (public API, orchestrator)
  server.ts      — Express + WS + routes (unchanged)

tests/
  client.test.ts      — renamed from tdlib-client.test.js
  fake-td-client.ts   — renamed from fake-td-client.js
  server.test.ts      — renamed from server.test.js
