--- src/utils/extract-text.ts
@@ -1,13 +1,0 @@
-interface TdMessageContent {
-  _: string;
-  text?: { text: string };
-  caption?: { text: string };
-  [key: string]: unknown;
-}
-
-export function extractText(content: TdMessageContent | null | undefined): string {
-  if (!content) return '';
-  if (content._ === 'messageText') return content.text!.text;
-  if (content.caption?.text) return content.caption.text;
-  return '';
-}
